Add withAuth, modified
There are two very important use-cases that were not currently satisfied
by `MonadAWS`:

- Getting the current credentials, e.g. for explicit passing to other
  tools that are not in this Haskell process
- Modifying `Env` in some way (timeout, logger, etc) for the duration of
  a block

This patch adds two members to the type class to solve this:

`withAuth` matches `Amazonka`'s `withAuth` but without the initial
argument, which is analagous to how `send`-et-al work. This seems like
an uncontroversial extension for that reason.

`modified` is a bit of a stretch, and I'm open to alternative ideas. For
any Reader-like instance, it's just `local`. I could name it `local` to
indicate that, but it's not necessarily the case and it would be a
constant collision, so I went with `modified`. `withEnv` is another
option, though I'd expect such a function to be `(Env -> m a) -> m a`,
which is not the functionality we need, nor what this provides.

In both cases, the `MockAWS` instance has to take some liberties, being
as it's a test instance without anything real going on.