Revert "fix(ci): Force CI to use specific nightly snapshot version"
This reverts commit 26622b971e90172a0d58e0e86bac32e3a8023246.

Looks like the stack-nightly.yaml is meant more as a placeholder to get
CI to try to build this package with `--resolver nightly`. This tends to
fail, but that's okay.

Moreover, since this file is meant more as a placeholder of sorts, we
don't need to include more info than what's necessary for this to be
considered a valid `stack.yaml` file. Thus, simply upgrading the
resolver version to the latest nightly version at the time of upgrading
will often suffice